automatically substitute hostname in JOSN info file
JSON info file provides name of server on status page of the node. Otherwise (or when you're outside of the freifunk network) the server shows up as an unknown neighbour. A special EUI64 address has to be provided by server for this to work! see commit a0ccf71540cb40be523fc1e543a284bf024e4a57.